ft: multiple buckets and mediator registration
