+++ .env.dev
@@ -17,7 +17,7 @@ BODY_SIZE_LIMIT=50mb
 MINIO_ENDPOINT=localhost
 MINIO_PORT=9000
 MINIO_USE_SSL=false
-MINIO_BUCKET=climate-mediator
+MINIO_BUCKETS=climate-mediator
 MINIO_ACCESS_KEY=tCroZpZ3usDUcvPM3QT6
 MINIO_SECRET_KEY=suVjMHUpVIGyWx8fFJHTiZiT88dHhKgVpzvYTOKK
 MINIO_PREFIX=
